Updated tests with new PaymentViewModel logic
